# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JobTracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Bad request
        "409":
          description: Email already in use
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Unauthorized
  /auth/otp/request:
    post:
      operationId: requestOtp
      tags: [auth]
      summary: Send a one-time email code (used for OTP login and password reset)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequestInput"
      responses:
        "200":
          description: Code sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpRequestResult"
        "400":
          description: Bad request
        "429":
          description: Too many requests — resend cooldown in effect
  /auth/otp/login:
    post:
      operationId: otpLogin
      tags: [auth]
      summary: Log in (or sign up) with an emailed one-time code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Bad request
        "401":
          description: Invalid or expired code
  /auth/password/reset:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Set a new password after verifying an emailed one-time code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetInput"
      responses:
        "200":
          description: Password updated
        "400":
          description: Bad request
        "401":
          description: Invalid or expired code
        "404":
          description: No account with that email
  /auth/verify/request:
    post:
      operationId: requestEmailVerification
      tags: [auth]
      summary: (Re)send a verification code to the current user's own email
      responses:
        "200":
          description: Code sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpRequestResult"
        "401":
          description: Unauthorized
        "429":
          description: Too many requests — resend cooldown in effect
  /auth/verify/confirm:
    post:
      operationId: confirmEmailVerification
      tags: [auth]
      summary: Confirm the current user's email with a one-time code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyConfirmInput"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Unauthorized, or invalid/expired code
...
  /auth/google:
    get:
      operationId: googleLogin
      tags: [auth]
      summary: Start the Google OAuth sign-in flow (browser redirect, not for XHR)
      responses:
        "302":
          description: Redirect to Google's consent screen
        "501":
          description: Google sign-in is not configured
  /auth/google/callback:
    get:
      operationId: googleCallback
      tags: [auth]
      summary: Google OAuth redirect target — completes login and redirects to the frontend
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirect back to the frontend, logged in (or with an ?error= query param)
        "501":
          description: Google sign-in is not configured
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [email, password]
      properties:
        name:
          type: ["string", "null"]
        email:
          type: string
        password:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [id, email, isAdmin, emailVerifiedAt, mustVerify, verifyGraceEndsAt]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        emailVerifiedAt:
          type: ["string", "null"]
        mustVerify:
          type: boolean
          description: True once the email-verification grace period has elapsed for an unverified account — the app must be fully blocked until verified.
        verifyGraceEndsAt:
          type: string
          description: ISO timestamp — when the grace period ends (createdAt + 48h) for an unverified account.
    OtpRequestInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
    OtpRequestResult:
      type: object
      required: [ok, resendIn]
      properties:
        ok:
          type: boolean
        resendIn:
          type: integer
          description: Seconds to wait before requesting another code.
        alreadyVerified:
          type: boolean
    OtpLoginInput:
      type: object
      required: [email, code]
      properties:
        email:
          type: string
        code:
          type: string
    PasswordResetInput:
      type: object
      required: [email, code, newPassword]
      properties:
        email:
          type: string
        code:
          type: string
        newPassword:
          type: string
    VerifyConfirmInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
...