# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JobTracker API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  id: string;
  email: string;
  /** @nullable */
  name?: string | null;
  isAdmin: boolean;
  /** @nullable */
  emailVerifiedAt: string | null;
  /** True once the email-verification grace period has elapsed for an unverified account — the app must be fully blocked until verified. */
  mustVerify: boolean;
  /** ISO timestamp — when the grace period ends (createdAt + 48h) for an unverified account. */
  verifyGraceEndsAt: string;
}