# lib/api-zod/src/generated/types/otpRequestResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JobTracker API
 * OpenAPI spec version: 0.1.0
 */

export interface OtpRequestResult {
  ok: boolean;
  /** Seconds to wait before requesting another code. */
  resendIn: number;
  alreadyVerified?: boolean;
}