# lib/api-zod/src/generated/types/passwordResetInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JobTracker API
 * OpenAPI spec version: 0.1.0
 */

export interface PasswordResetInput {
  email: string;
  code: string;
  newPassword: string;
}